Maybe works double passthorugh

+++ cl-test/conv_old.cl
@@ -1,12 +1,12 @@
-void pixconv(float *acc, float kern, global const uchar *img, int pos) {
+void pixconv(double *acc, double kern, global const uchar *img, int pos) {
     acc[0] += kern * (float)convert_float(img[pos*3]);
     acc[1] += kern * (float)convert_float(img[pos*3+1]);
     acc[2] += kern * (float)convert_float(img[pos*3+2]);
 }
 
-void kernel conv(global const float *convKern, const int convWidth, global const uchar *image, global uchar *outimg, const int w, const int h, global int* C) {
+void kernel conv(global const double *convKern, const int convWidth, global const uchar *image, global uchar *outimg, const int w, const int h, global int* C) {
   int id = get_global_id(0);
-  float acc[3] = {0,0,0};
+  double acc[3] = {0,0,0};
   int x = id % w;
   int y = id / w;
   int f = 255;
@@ -23,7 +23,7 @@ void kernel conv(global const float *convKern, const int convWidth, global const
     }
   }
 
-  outimg[id*3] = convert_uchar(clamp(acc[0], 0.0f, 255.0f));
-  outimg[id*3+1] = convert_uchar(clamp(acc[1], 0.0f, 255.0f));
-  outimg[id*3+2] = convert_uchar(clamp(acc[2], 0.0f, 255.0f));
+  outimg[id*3] = convert_uchar(clamp(acc[0], 0.0, 255.0));
+  outimg[id*3+1] = convert_uchar(clamp(acc[1], 0.0, 255.0));
+  outimg[id*3+2] = convert_uchar(clamp(acc[2], 0.0, 255.0));
 }
